Version 1.8.77 Updates - tuning, maintenance New features: Page Manage Schema
1.Export UI Defaults - download an UI Defaults definition PL/SQL script file for tables of the current schema with addition attributes including width, max length, format masks, and LOVs for references, Yes/No fields and check constraints with IN-Lists.
You can Import this file with APEX to improve the default appearance of new forms on your tables.

2. Export Updatable Views - download an PL/SQL Script with Create View statements for updatable views.
The data browser can display this views  'Import View'.
The views will have INSTEAD OF INSERT OR UPDATe OR DELETE triggers that will manage the processing of DML statements on that views.
You can Import this file with APEX SQL Workshop to install this views.